refactor(server): introduce TerminalBackend spine; dissolve meta/* orchestrators (R-1)
Carve the kolu-server at the volatility axis "where a terminal lives": one
`TerminalBackend` interface in `kolu-common/terminalBackend` with the only
operations that change between local and (future) remote variants — PTY
spawn, kill, per-terminal channel subscribe, fs surface, git surface.

`LocalTerminalBackend` (the sole concrete shape today) absorbs every
per-terminal orchestrator the server used to split across `meta/*.ts`:
agent-command tracker, agent detectors for claude-code/codex/opencode,
git resolver, github PR watcher, foreground process observer. They live
as private functions inside the file because their lifecycle is owned by
`spawnPty` — no other call site, and splitting them across files only
forced multiple `cwd:<id>`/`commandRun:<id>`/`title:<id>` subscribers to
coordinate via the publisher when they could just call each other.

Every consumer downstream — router, surface streams, terminal lifecycle —
resolves through `getTerminalBackendFor(location)` (R-1: only the local
variant; R-2 will add the resolver branch on `host`). The sole place that
pattern-matches on `location.kind` is the resolver itself.

No user-visible change. Behavior preserved because `LocalTerminalBackend`
invokes the same integration-package functions (`kolu-claude-code`,
`kolu-git`, …) the orchestrators called; existing test suite covers
the publish-routing fence (`updateServerMetadata` fires `terminals:dirty`;
`updateServerLiveMetadata` does not) and the recency-bump pure function.
