+++ packages/common/package.json
@@ -7,6 +7,7 @@
   "sideEffects": false,
   "exports": {
     "./surface": "./src/surface.ts",
+    "./terminalBackend": "./src/terminalBackend.ts",
     "./contract": "./src/contract.ts",
     "./errors": "./src/errors.ts",
     "./config": "./src/config.ts",
