+++ packages/server/src/terminal-registry.ts
@@ -2,16 +2,12 @@
  * Terminal registry — the `Map<TerminalId, TerminalProcess>` and the
  * pure read/write accessors around it.
  *
- * Kept as a leaf module on purpose: metadata providers under `./meta/*`
- * need to read the registry (for `TerminalProcess` shape and for
- * `getTerminal`), and the higher-level lifecycle in `./terminals.ts` needs
- * to write to it. When both lived on `terminals.ts`, the providers' edge
- * back to `terminals.ts` (plus `terminals.ts`'s edge to `./meta/index.ts`
- * for `startProviders`) closed a cycle that Biome's `noImportCycles`
- * correctly flagged (#710). Splitting the map+reads here breaks it.
- *
- * No imports from `./meta/*` or `./terminals.ts` — that invariant is what
- * keeps this file a leaf.
+ * Backend-agnostic: every `TerminalBackend` (local in R-1, remote in
+ * R-2) writes to the same registry so consumers downstream (router,
+ * surface) iterate one place regardless of where the terminal lives.
+ * Per-backend internal state (PTY handle, provider cleanups for
+ * `LocalTerminalBackend`) stays inside the backend itself, not on
+ * `TerminalProcess`.
  */
 
 import type {
@@ -19,18 +15,19 @@ import type {
   TerminalInfo,
   TerminalMetadata,
 } from "kolu-common/surface";
-import type { PtyHandle } from "kolu-pty";
+import type { TerminalHandle } from "kolu-common/terminalBackend";
 
 /** Server-side terminal state. `info` is the wire shape sent in the
- *  `terminalList` cell snapshot; `meta` is mutated in place by providers
- *  and published via the `terminalMetadata` collection from
- *  `meta/state.ts`. */
+ *  `terminalList` cell snapshot; `meta` is mutated in place by the
+ *  owning backend's providers and published via the
+ *  `terminalMetadata` collection from `terminalBackend/metadata.ts`;
+ *  `handle` is the abstract control surface (write / resize / screen
+ *  state — NO `dispose()`, the backend's `killTerminal` is the sole
+ *  termination path). */
 export interface TerminalProcess {
   info: TerminalInfo;
   meta: TerminalMetadata;
-  handle: PtyHandle;
-  /** Cleanup function for all metadata providers. */
-  stopProviders: () => void;
+  handle: TerminalHandle;
 }
 
 const terminals = new Map<TerminalId, TerminalProcess>();
@@ -45,11 +42,11 @@ export function unregisterTerminal(id: TerminalId): boolean {
   return terminals.delete(id);
 }
 
-/** Snapshot + clear. Used by `killAllTerminals` where the caller needs to
- *  dispose each handle AFTER the map is empty (so onExit callbacks can't
- *  find the entry and trigger session saves). Returning the entries keeps
- *  the clear-then-dispose ordering in the caller rather than forcing it
- *  into the registry API. */
+/** Snapshot + clear. Used by `killAllTerminals` where the caller needs
+ *  to dispose each handle AFTER the map is empty (so onExit callbacks
+ *  can't find the entry and trigger session saves). Returning the
+ *  entries keeps the clear-then-dispose ordering in the caller rather
+ *  than forcing it into the registry API. */
 export function drainTerminals(): TerminalProcess[] {
   const entries = [...terminals.values()];
   terminals.clear();
@@ -67,8 +64,8 @@ export function terminalEntries(): IterableIterator<
 /** Current terminals in their canonical `Map` insertion order.
  *
  *  Insertion order is the ordering model — new terminals append to the
- *  tail. Clients render this order directly; within-group pill ordering
- *  is a separate spatial sort driven by saved canvas layouts. */
+ *  tail. Clients render this order directly; within-group pill
+ *  ordering is a separate spatial sort driven by saved canvas layouts. */
 export function listTerminals(): TerminalInfo[] {
   return [...terminals.values()].map((entry) => entry.info);
 }
@@ -77,9 +74,10 @@ export function listTerminals(): TerminalInfo[] {
 export const terminalCount = (): number => terminals.size;
 
 /** Number of terminals currently hosting a Claude Code session. Derived
- *  from `entry.meta.agent` — the generic agent orchestrator
- *  (`meta/agent.ts`, driven by `claudeCodeProvider` from `kolu-claude-code`)
- *  sets it on session match and clears it on teardown. Exported for diagnostics. */
+ *  from `entry.meta.agent` — the agent detectors inside
+ *  `LocalTerminalBackend` (driven by `claudeCodeProvider` from
+ *  `kolu-claude-code`) set it on session match and clear it on
+ *  teardown. Exported for diagnostics. */
 export function countActiveClaudeSessions(): number {
   let n = 0;
   for (const entry of terminals.values()) {
