+++ packages/server/src/terminals.ts
@@ -1,14 +1,19 @@
 /**
- * Terminal lifecycle: spawn PTYs, wire them to metadata providers, and
- * manage create/kill/update operations. The underlying `Map` and its
- * simple accessors (`getTerminal`, `listTerminals`, `terminalCount`,
- * `countActiveClaudeSessions`, the `TerminalProcess` shape) live in
- * `./terminal-registry.ts` so `./meta/*` can depend on the registry
- * without closing a cycle back through this file.
+ * Terminal lifecycle façade — `createTerminal` / `killTerminal` /
+ * `killAllTerminals` resolve to a `TerminalBackend` via
+ * `getTerminalBackendFor(location)` and delegate. The backend owns
+ * PTY spawn, per-terminal provider startup, registry insert/remove,
+ * autosave-trigger signalling.
  *
- * External callers that used to import state-reads + lifecycle from
- * `./terminals.ts` as a single module keep their import path — this
- * file re-exports the registry surface they need.
+ * Client-facing per-terminal metadata setters (`setTerminalParent`,
+ * `setCanvasLayout`, `setSubPanelState`, `setRightPanelState`,
+ * `setTerminalTheme`, `setTerminalIntent`) live here because they're
+ * location-agnostic — they mutate the in-registry entry through the
+ * narrowed `updateClientMetadata` helper, which publishes through the
+ * same metadata channel regardless of which backend owns the terminal.
+ *
+ * Re-exports the registry surface for callers that used to import
+ * state-reads + lifecycle from this file as a single module.
  */
 
 import type {
@@ -18,29 +23,10 @@ import type {
   TerminalId,
   TerminalInfo,
 } from "kolu-common/surface";
-import { DEFAULT_SCROLLBACK } from "kolu-common/config";
-import { spawnPty } from "kolu-pty";
-import pkg from "../package.json" with { type: "json" };
-import { cleanupTerminalScratch } from "./terminalScratch.ts";
-import { koluShellDir } from "./koluRoot.ts";
-import { log } from "./log.ts";
-import {
-  createMetadata,
-  startProviders,
-  updateClientMetadata,
-  updateServerMetadata,
-} from "./meta/index.ts";
-import { terminalChannels, terminalsDirtyChannel } from "./publisher.ts";
-import { surfaceCtx } from "./surface.ts";
-import {
-  drainTerminals,
-  getTerminal,
-  listTerminals,
-  registerTerminal,
-  type TerminalProcess,
-  terminalEntries,
-  unregisterTerminal,
-} from "./terminal-registry.ts";
+import { updateClientMetadata } from "./terminalBackend/metadata.ts";
+import { getTerminalBackendFor } from "./terminalBackend/index.ts";
+import { terminalsDirtyChannel } from "./publisher.ts";
+import { getTerminal, terminalEntries } from "./terminal-registry.ts";
 
 // Re-export registry accessors + type so external callers (router.ts,
 // diagnostics.ts, index.ts) keep a single import path.
@@ -78,124 +64,29 @@ export function snapshotSession(): {
   return { terminals: snappedTerminals, activeTerminalId };
 }
 
-/** Notify that terminal state changed (triggers debounced session auto-save). */
-function emitChanged(): void {
-  terminalsDirtyChannel.publish({});
-}
-
-/** Notify that terminal membership changed (create/kill).
- *  Drives the live `surface.terminalList.get` stream to clients. The
- *  surface owns the publish channel; calling `set` triggers the
- *  framework's apply+publish chain (the `terminalList` cell's store is a
- *  no-op since the registry is canonical). */
-function emitListChanged(): void {
-  surfaceCtx.cells.terminalList.set(listTerminals());
-}
-
-/** Create a new terminal, spawn a PTY process. `initial` seeds
- *  client-owned metadata before `startProviders` runs, so the first
- *  `terminalMetadata` collection read carries it — used by session
- *  restore to avoid racing post-hoc `setCanvasLayout` / `setTheme` /
- *  `setSubPanel` RPCs against the client's canvas-cascade effect (#642). */
+/** Create a new terminal. The backend owns PTY spawn, provider
+ *  startup, and registry insert; this wrapper just resolves the
+ *  backend, mints an id, and forwards. `initial` seeds client-owned
+ *  metadata before providers run — see #642 (avoids racing post-hoc
+ *  `setCanvasLayout` / `setTheme` / `setSubPanel` RPCs against the
+ *  client's canvas-cascade effect). */
 export function createTerminal(
   cwd?: string,
   parentId?: string,
   initial?: InitialTerminalMetadata,
 ): TerminalInfo {
   const id = crypto.randomUUID();
-  const tlog = log.child({ terminal: id });
-
-  const handle = spawnPty(
-    tlog,
-    id,
-    {
-      rcDir: koluShellDir,
-      termProgramVersion: pkg.version,
-      scrollback: DEFAULT_SCROLLBACK,
-      onData: (data) => {
-        terminalChannels.data(id).publish(data);
-      },
-      // On natural exit: notify clients, then remove from server state
-      onExit: (exitCode) => {
-        tlog.info({ exitCode }, "exited");
-        const entry = getTerminal(id);
-        if (entry) {
-          entry.stopProviders();
-          cleanupTerminalScratch(id);
-        }
-        surfaceCtx.events.terminalExit.publish({ id }, exitCode);
-        // Only save session on natural exit (entry still in map).
-        // killAllTerminals clears the map first, so entry is gone — skip.
-        const wasNaturalExit = unregisterTerminal(id);
-        if (wasNaturalExit) {
-          emitChanged();
-          emitListChanged();
-        }
-      },
-      // PTY callback (OSC 0/2): notify process provider that title changed
-      onTitleChange: (title) => {
-        terminalChannels.title(id).publish(title);
-      },
-      // PTY callback (OSC 633;E): raw preexec command line. Agent parsing,
-      // the per-terminal stash, and the recent-agents MRU all live in
-      // `meta/agent-command.ts`, fed via this channel.
-      onCommandRun: (raw) => {
-        terminalChannels.commandRun(id).publish(raw);
-      },
-      // PTY callback (OSC 7): update metadata CWD, notify providers via cwd channel
-      onCwd: (newCwd) => {
-        const entry = getTerminal(id);
-        if (entry) {
-          updateServerMetadata(entry, id, (m) => {
-            m.cwd = newCwd;
-          });
-          terminalChannels.cwd(id).publish(newCwd);
-        }
-      },
-    },
-    cwd,
-  );
-
-  const meta = createMetadata(handle.cwd);
-  if (parentId) meta.parentId = parentId;
-  // Seed client-owned initial metadata BEFORE startProviders so the first
-  // `terminalMetadata` collection yield carries these fields (see #642).
-  if (initial?.themeName) meta.themeName = initial.themeName;
-  if (initial?.canvasLayout) meta.canvasLayout = initial.canvasLayout;
-  if (initial?.subPanel) meta.subPanel = initial.subPanel;
-  if (initial?.rightPanel) meta.rightPanel = initial.rightPanel;
-  if (initial?.lastActivityAt !== undefined)
-    meta.lastActivityAt = initial.lastActivityAt;
-  if (initial?.intent) meta.intent = initial.intent;
-  const entry: TerminalProcess = {
-    info: { id, pid: handle.pid },
-    meta,
-    handle,
-    stopProviders: () => {},
-  };
-  // Start providers after entry is in the map (providers may emit immediately)
-  registerTerminal(id, entry);
-  entry.stopProviders = startProviders(entry, id);
-
-  tlog.info({ pid: handle.pid, total: listTerminals().length }, "created");
-  emitChanged();
-  emitListChanged();
-  return entry.info;
+  // R-1: all terminals are local; R-2's `getTerminalBackendForCreate`
+  // resolver will read `parentId` to inherit the parent's location.
+  const backend = getTerminalBackendFor({ kind: "local" });
+  return backend.spawnPty(id, { cwd, parentId, initialMetadata: initial });
 }
 
-/** Kill a terminal's PTY process and remove it from the map. Returns final info, or undefined if not found. */
+/** Kill a terminal. Returns final info, or undefined if not found. */
 export function killTerminal(id: TerminalId): TerminalInfo | undefined {
   const entry = getTerminal(id);
   if (!entry) return undefined;
-
-  log.child({ terminal: id }).info({ pid: entry.handle.pid }, "killing");
-  entry.stopProviders();
-  entry.handle.dispose();
-  cleanupTerminalScratch(id);
-  unregisterTerminal(id);
-  emitChanged();
-  emitListChanged();
-  return entry.info;
+  return getTerminalBackendFor({ kind: "local" }).killTerminal(id);
 }
 
 /** Set or clear a terminal's parent relationship. */
@@ -234,7 +125,8 @@ export function setCanvasLayout(
  *  Equality-gated: the client RPCs this on every drag tick of the
  *  resizable handle, so without a guard each mouse-move would fan a
  *  full per-key metadata publish to every connected client. Same shape
- *  as `meta/agent-command.ts`'s `lastAgentCommand` gate. */
+ *  as the `lastAgentCommand` gate inside `LocalTerminalBackend`'s
+ *  agent-command tracker. */
 export function setSubPanelState(
   id: TerminalId,
   state: { collapsed: boolean; panelSize: number },
@@ -258,9 +150,9 @@ export function setSubPanelState(
  *  refresh) pick up the change from the same channel as every other
  *  client-owned metadata field.
  *
- *  Equality-gated like `setSubPanelState` — the client RPCs this on every
- *  file-tree click and tab-toggle, so without a guard each interaction
- *  would fan a full per-key metadata publish. Deep-compares
+ *  Equality-gated like `setSubPanelState` — the client RPCs this on
+ *  every file-tree click and tab-toggle, so without a guard each
+ *  interaction would fan a full per-key metadata publish. Deep-compares
  *  `selectedFileByMode` since the user clicks files often. */
 export function setRightPanelState(
   id: TerminalId,
@@ -295,12 +187,12 @@ let activeTerminalId: TerminalId | null = null;
 
 /** Store which terminal is active (reported by the client).
  *  Only emits session:changed when a terminal is actually selected —
- *  null (no selection, e.g. client reconnect) must not trigger auto-save
- *  because snapshotSession() may return an empty terminal list at that
- *  point, which would clear the saved session. */
+ *  null (no selection, e.g. client reconnect) must not trigger
+ *  auto-save because snapshotSession() may return an empty terminal
+ *  list at that point, which would clear the saved session. */
 export function setActiveTerminalId(id: TerminalId | null): void {
   activeTerminalId = id;
-  if (id !== null) emitChanged();
+  if (id !== null) terminalsDirtyChannel.publish({});
 }
 
 /** Set the theme name for a terminal (stored in metadata, published to clients). */
@@ -325,14 +217,5 @@ export function setTerminalIntent(id: TerminalId, intent: string): void {
 
 /** Kill and remove all terminals. Used by tests to reset server state between scenarios. */
 export function killAllTerminals(): void {
-  // Snapshot entries and clear map BEFORE disposing — prevents onExit
-  // callbacks from finding terminals and triggering session saves.
-  const entries = drainTerminals();
-  log.info({ count: entries.length }, "killing all terminals");
-  for (const entry of entries) {
-    entry.stopProviders();
-    entry.handle.dispose();
-    cleanupTerminalScratch(entry.info.id);
-  }
-  emitListChanged();
+  getTerminalBackendFor({ kind: "local" }).killAllTerminals();
 }
